updating db search templates

+++ trunk/OGen-Kick/src/Kick-businesslayer-shared/_base/IBO0_CRD_Permition.cs
@@ -25,7 +25,7 @@ namespace OGen.NTier.Kick.lib.businesslayer.shared {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		);
 	}
