+++ trunk/OGen-Kick/src/Kick-businesslayer-shared/_base/IBO0_NWS_Attachment.cs
@@ -37,7 +37,7 @@ namespace OGen.NTier.Kick.lib.businesslayer.shared {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		);
 		OGen.NTier.Kick.lib.datalayer.shared.structures.SO_vNWS_Attachment[] getRecord_byContent_andLanguage(
@@ -48,7 +48,7 @@ namespace OGen.NTier.Kick.lib.businesslayer.shared {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		);
 		long insObject(
