+++ trunk/OGen-Kick/src/Kick-businesslayer/_base/BO0_NWS_Attachment.cs
@@ -59,7 +59,7 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return OGen.NTier.Kick.lib.businesslayer.SBO_NWS_Attachment.getRecord_byContent(
@@ -83,7 +83,7 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return OGen.NTier.Kick.lib.businesslayer.SBO_NWS_Attachment.getRecord_byContent_andLanguage(
