+++ trunk/OGen-Kick/src/Kick-businesslayer/_base/BO0_WEB_User.cs
@@ -63,7 +63,7 @@ namespace OGen.NTier.Kick.lib.businesslayer {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return OGen.NTier.Kick.lib.businesslayer.SBO_WEB_User.getRecord_generic(
