+++ trunk/OGen-Kick/src/Kick-remoting-client/_base/RC0_DIC_Dic.cs
@@ -64,7 +64,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return bo_.getRecord_byApplication(
@@ -86,7 +86,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return bo_.getRecord_byLanguage(
@@ -109,7 +109,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return bo_.getRecord_Language(
