+++ trunk/OGen-Kick/src/Kick-remoting-client/_base/RC0_NWS_News.cs
@@ -82,7 +82,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return bo_.getRecord_byContent(
@@ -116,7 +116,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return bo_.getRecord_generic(
