+++ trunk/OGen-Kick/src/Kick-remoting-client/_base/RC0_WEB_DefaultProfile.cs
@@ -49,7 +49,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return bo_.getRecord_all(
