+++ trunk/OGen-Kick/src/Kick-remoting-client/_base/RS0_CRD_Profile.cs
@@ -54,7 +54,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.server {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			throw new Exception("your not calling the remoting server, but the client's remoting server implementation");
@@ -68,7 +68,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.server {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			throw new Exception("your not calling the remoting server, but the client's remoting server implementation");
@@ -82,7 +82,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.server {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			throw new Exception("your not calling the remoting server, but the client's remoting server implementation");
@@ -96,7 +96,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.server {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			throw new Exception("your not calling the remoting server, but the client's remoting server implementation");
