+++ trunk/OGen-Kick/src/Kick-remoting-client/_base/RS0_NWS_Attachment.cs
@@ -54,7 +54,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.server {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			throw new Exception("your not calling the remoting server, but the client's remoting server implementation");
@@ -69,7 +69,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.server {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			throw new Exception("your not calling the remoting server, but the client's remoting server implementation");
