+++ trunk/OGen-Kick/src/Kick-remoting-server/_base/RS0_WEB_User.cs
@@ -73,7 +73,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.remoting.server {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return OGen.NTier.Kick.lib.businesslayer.SBO_WEB_User.getRecord_generic(
