+++ trunk/OGen-Kick/src/Kick-webservices-client/_base/WC0_CRD_Profile.cs
@@ -120,7 +120,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -134,7 +134,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_CRD_Profile[])results[0];
 		}
@@ -154,7 +154,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -168,7 +168,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_CRD_ProfileProfile[])results[0];
 		}
@@ -188,7 +188,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -202,7 +202,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_vCRD_ProfilePermition[])results[0];
 		}
@@ -222,7 +222,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -236,7 +236,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_vCRD_UserProfile[])results[0];
 		}
