+++ trunk/OGen-Kick/src/Kick-webservices-client/_base/WC0_NWS_News.cs
@@ -122,7 +122,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -136,7 +136,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_vNWS_Content[])results[0];
 		}
@@ -167,7 +167,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -192,7 +192,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_vNWS_Content[])results[0];
 		}
