+++ trunk/OGen-Kick/src/Kick-webservices-client/_base/WC0_NWS_Profile.cs
@@ -119,7 +119,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -132,7 +132,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_vNWS_Profile[])results[0];
 		}
@@ -151,7 +151,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -164,7 +164,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_vNWS_Profile[])results[0];
 		}
@@ -184,7 +184,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -198,7 +198,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_NWS_ContentProfile[])results[0];
 		}
