+++ trunk/OGen-Kick/src/Kick-webservices-client/_base/WC0_NWS_Tag.cs
@@ -148,7 +148,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -162,7 +162,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_vNWS_Tag[])results[0];
 		}
@@ -182,7 +182,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -196,7 +196,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_NWS_ContentTag[])results[0];
 		}
@@ -216,7 +216,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -230,7 +230,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_vNWS_Tag[])results[0];
 		}
@@ -250,7 +250,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			object[] results = this.Invoke(
@@ -264,7 +264,7 @@ namespace OGen.NTier.Kick.lib.distributedlayer.webservices.client {
 					page_numRecords_in
 				}
 			);
-			page_itemsCount_out = (int)results[1];
+			page_itemsCount_out = (long)results[1];
 			errors_out = (System.Int32[])results[2];
 			return (OGen.NTier.Kick.lib.datalayer.shared.structures.SO_vNWS_Tag[])results[0];
 		}
