+++ trunk/OGen-Kick/src/Kick-webservices-server/_base/WS0_NWS_News.cs
@@ -76,7 +76,7 @@ namespace OGen.NTier.Kick.distributedlayer.webservices.server {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return OGen.NTier.Kick.lib.businesslayer.SBO_NWS_News.getRecord_byContent(
@@ -113,7 +113,7 @@ namespace OGen.NTier.Kick.distributedlayer.webservices.server {
 			int page_orderBy_in, 
 			int page_in, 
 			int page_numRecords_in, 
-			out int page_itemsCount_out, 
+			out long page_itemsCount_out, 
 			out System.Int32[] errors_out
 		) {
 			return OGen.NTier.Kick.lib.businesslayer.SBO_NWS_News.getRecord_generic(
